requirements.txt
django
channels
daphne
django-crispy-forms
crispy-bootstrap4
django-autoslug
pillow